Changes to build ARM64
These changes allow the code to build with Rack-ARM api and run in
rack free for ARM on my M1.

They really are just sybbing out the direct includes of the intrinsics
with rack.hpp but see the comment in arm_intrinsics_sub for a different
stragegy you may want.